+++ src/main/java/pl/asie/computronics/cc/IntegrationComputerCraft.java
@@ -11,8 +11,9 @@ import pl.asie.computronics.integration.appeng.DriverSpatialIOPort;
 import pl.asie.computronics.integration.buildcraft.DriverHeatable;
 import pl.asie.computronics.integration.cofh.DriverEnergyHandler;
 import pl.asie.computronics.integration.enderio.DriverAbstractMachine;
-import pl.asie.computronics.integration.enderio.DriverCapacitorBankOld;
+import pl.asie.computronics.integration.enderio.DriverAbstractPoweredMachine;
 import pl.asie.computronics.integration.enderio.DriverCapacitorBank;
+import pl.asie.computronics.integration.enderio.DriverCapacitorBankOld;
 import pl.asie.computronics.integration.enderio.DriverHasExperience;
 import pl.asie.computronics.integration.enderio.DriverIOConfigurable;
 import pl.asie.computronics.integration.enderio.DriverPowerStorage;
@@ -112,6 +113,7 @@ public class IntegrationComputerCraft {
 				registerMultiPeripheralProvider(new DriverHasExperience.CCDriver());
 				registerMultiPeripheralProvider(new DriverPowerStorage.CCDriver());
 				registerMultiPeripheralProvider(new DriverAbstractMachine.CCDriver());
+				registerMultiPeripheralProvider(new DriverAbstractPoweredMachine.CCDriver());
 				registerMultiPeripheralProvider(new DriverCapacitorBank.CCDriver());
 				registerMultiPeripheralProvider(new DriverCapacitorBankOld.CCDriver());
 				registerMultiPeripheralProvider(new DriverTransceiver.CCDriver());