+++ src/main/java/pl/asie/computronics/integration/waila/providers/WailaLocomotiveRelay.java
@@ -1,6 +1,5 @@
 package pl.asie.computronics.integration.waila.providers;
 
-import cpw.mods.fml.common.Loader;
 import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaDataAccessor;
 import net.minecraft.item.ItemStack;
@@ -8,7 +7,6 @@ import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
 import pl.asie.computronics.integration.waila.ConfigValues;
-import pl.asie.computronics.reference.Mods;
 import pl.asie.computronics.tile.TileLocomotiveRelay;
 import pl.asie.computronics.util.StringUtil;