+++ src/main/java/pl/asie/computronics/oc/IntegrationOpenComputers.java
@@ -16,8 +16,9 @@ import pl.asie.computronics.integration.betterstorage.DriverCrateStorageNew;
 import pl.asie.computronics.integration.betterstorage.DriverCrateStorageOld;
 import pl.asie.computronics.integration.buildcraft.DriverHeatable;
 import pl.asie.computronics.integration.enderio.DriverAbstractMachine;
-import pl.asie.computronics.integration.enderio.DriverCapacitorBankOld;
+import pl.asie.computronics.integration.enderio.DriverAbstractPoweredMachine;
 import pl.asie.computronics.integration.enderio.DriverCapacitorBank;
+import pl.asie.computronics.integration.enderio.DriverCapacitorBankOld;
 import pl.asie.computronics.integration.enderio.DriverHasExperience;
 import pl.asie.computronics.integration.enderio.DriverIOConfigurable;
 import pl.asie.computronics.integration.enderio.DriverPowerStorage;
@@ -168,6 +169,7 @@ public class IntegrationOpenComputers {
 				Driver.add(new DriverHasExperience.OCDriver());
 				Driver.add(new DriverPowerStorage.OCDriver());
 				Driver.add(new DriverAbstractMachine.OCDriver());
+				Driver.add(new DriverAbstractPoweredMachine.OCDriver());
 				Driver.add(new DriverCapacitorBank.OCDriver());
 				Driver.add(new DriverCapacitorBankOld.OCDriver());
 				Driver.add(new DriverTransceiver.OCDriver());