+++ src/main/java/pl/asie/computronics/tile/TileLocomotiveRelay.java
@@ -1,6 +1,5 @@
 package pl.asie.computronics.tile;
 
-import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.Optional;
 import dan200.computercraft.api.lua.ILuaContext;
 import dan200.computercraft.api.lua.LuaException;