+++ changelog.md
@@ -19,6 +19,7 @@
     - New Capacitor Banks can be properly accessed by CC and OC now!
     - for CC, getEnergyStored()/getMaxEnergyStored() has been fixed
     - for OC, you will temporarily have to use getEnergyStoredForNetwork() and getMaxEnergyStoredForNetwork() until a bug in OC gets fixed. Once that has been fixed, you will be able to use getEnergyStored() and getMaxEnergyStored() as well.
+    - Also fixed getPowerPerTick() for powerable EnderIO devices
 
 ==== 1.3.2 ====